libmctp: add debug token install/query/erase commands
The commit will also implment several improvements except for adding new
commands.
1. add the verbose option to suspress the message from
vdm_resp_output() and print_hex()
2. using ENCODE_VMD_CMD_FUNC to avoid the duplicate code.

Tested:
Using mctp-vdm-util with debug_token_query/erase/install and the
responses will be returned.

Usages:
./mctp-vdm-util -c command_options -t eid
options: debug_token_query/debug_token_install/debug_token_erase

Fixes Jira https://jirasw.nvidia.com/browse/DGXOPENBMC-4958
Fixes Jira https://jirasw.nvidia.com/browse/DGXOPENBMC-4959
Fixes Jira https://jirasw.nvidia.com/browse/DGXOPENBMC-4960

+++ ctrld/mctp-socket.c
@@ -96,7 +96,7 @@ mctp_requester_rc_t mctp_client_recv(mctp_eid_t eid, int mctp_fd,
 	size_t min_len = sizeof(eid) + sizeof(MCTP_MSG_TYPE_HDR) +
 			 sizeof(struct mctp_ctrl_cmd_msg_hdr);
 
-	int length = recv(mctp_fd, NULL, 0, MSG_PEEK | MSG_TRUNC);
+	ssize_t length = recv(mctp_fd, NULL, 0, MSG_PEEK | MSG_TRUNC);
 
 	if (length < 0 && errno == EAGAIN) {
 		MCTP_CTRL_INFO("%s: Recv failed: due to timedout\n", __func__);
